feat: add "GraalVM" plugin to create native executable
