+++ app/src/main/resources/META-INF/native-image/predefined-classes-config.json
@@ -1,0 +1,7 @@
+[
+  {
+    "type":"agent-extracted",
+    "classes":[
+    ]
+  }
+]
