+++ app/src/main/resources/META-INF/native-image/proxy-config.json
@@ -1,0 +1,2 @@
+[
+]
