+++ app/src/main/resources/META-INF/native-image/resource-config.json
@@ -1,0 +1,10 @@
+{
+  "resources":{
+  "includes":[{
+    "pattern":"\\QMETA-INF/services/java.sql.Driver\\E"
+  }]},
+  "bundles":[{
+    "name":"jline.console.completer.CandidateListCompletionHandler",
+    "locales":["und"]
+  }]
+}
